Add recursive check for substeps.
If a QuestStep has substeps, the correct top-level step is not
highlighted in the corresponding panel if the player is currently doing
a substep. This change adds .contains() to recursively search the
substeps of a step instead of searching through the steps explicitly
added to a panel.

All credit goes to @Senmori.