Add quest rewards to side-panel
Added all quest rewards to the side-panel, as has been requested a few times.

Functions the same as other recommended and required side-panels.

Panels are set to always show, as each quest should have a reward by default. (May need to be changed if Custom quests are added in the future). This is done so that a report can easily be made if a quest is missing its rewards.