Resolving Wanted! essence issue #478
Issue #478 explains that the guide states the user has all required items if they have some combination of rune/pure essence that are unnoted and amount to something >= 20. The quest needs all 20 essence to be the same (either all rune or all pure essence), so I split the logic into an OR, much like the commorb vs 10k requirement.