Update TheSlugMenace.java Quest Step Description
Update quest step description when crafting runes to include the following line: 
"Bring extra essence (~10 extra) as it is possible to accidentally destroy the essence upon creation.". 

The "10 extra" suggestion comes from the original item requirement section, where 5 runes are required, but it is recommended to bring 15 to be safe. The "15 to be safe" may be forgotten by the time that step is reached, so a reminder is helpful. 

Old:
"Right-click each page to turn rune/pure essence into empty runes. Take each empty rune and use it on its respective Runecrafting Altar. Bring extra essence (~10 extra) as it is possible to accidentally destroy the essence upon creation."

New: 
"Right-click each page to turn rune/pure essence into empty runes. Take each empty rune and use it on its respective Runecrafting Altar. Bring extra essence (~10 extra) as it is possible to accidentally destroy the essence upon creation."