Item wiki redirect
Set the url suffix of "food" to send the user to the "Food" section of the wiki when opening the wiki page from the item. This page lists the different foods. Without this addition a page of one of the fish in the food ItemCollection is linked to.

The same is done for "axe". As ItemCollection is used the wiki page that opens is "crystal axe", the suffix redirects to the list of axes.

This commit acts as an example for how the setUrlSuffix() method can be used.