Feature: Item -> Go to Wiki.. (#889)
* Feature: Item -> Go to Wiki..

* Methods for wiki url suffix

Getter/setter for the Wiki URL suffix used in the Item -> Go to WIki..

* update copy method

Updated copy method to include suffix settings (if some had been set) for an item. This makes it so these would not be lost when copy() is called on the item.

* Popup and link logic

Popup panel and link therein added to the so right clicking an item opens the popup which houses the single option "go to wiki.." which sends you to the items wiki page

* Item wiki redirect

Set the url suffix of "food" to send the user to the "Food" section of the wiki when opening the wiki page from the item. This page lists the different foods. Without this addition a page of one of the fish in the food ItemCollection is linked to.

The same is done for "axe". As ItemCollection is used the wiki page that opens is "crystal axe", the suffix redirects to the list of axes.

This commit acts as an example for how the setUrlSuffix() method can be used.

* Convert ItemCollections to enum

* Update FollowerItemRequirement.java

* Light Sources collection modification

"Light Sources" would default to firemaking cape, modificaiton changes this to the "Light sources" wiki page

Co-authored-by: Zoinkwiz <zoinkwiz@hotmail.co.uk>