Changes to the Seers' course
Changes to the Seers' course and use this as a base template to work on other courses etc. If anything else you could try change how it looks and stuff eventually but for now it is fine!

Keeps a small box with info and sidebar where to start course without being cluttered.