feat: Tele step to Lord Hosidius
Again keeping it simple since inventory space can be annoying here, and users haven't been prompted to bring a Xeric's
Talisman.

In a bigger rewrite, asking the user to bring a Xeric's Talisman too might be worth it, and it could then be worth
revisiting some of the teleport suggestions for better alternatives.
In that alternative, it might be good to point out banking opportunities that flow well with the quest.