chore: Remove unused PlayerModelRequirement class
It was previously used for the "The Recruitment Drive" quest, but hasn't been used for quite a while now.
Since the class is very specifically made for the female/male model check, which is deprecated in RuneLite, I think
this class can be removed & re-implemented if any other model requirements are added in the future.