Lumbridge & Draynor Hard Diary: Polish Power Amulet step
The "Move to the Lumbridge furnace" has been changed to an ObjectStep which just points at the correct furnace

The step coords were wrong, and were outside of the area where it would tell you to actually smelt the diamond amulet.
Alternatively, the step could've been removed entirely but this felt like an easy enough fix.
I also changed the wording of it to say "Furnace" instead of "Smithy"

The 3-step-process of smelting -> stringing -> enchanting were flipped, so it checks the last step first.
Previously it would get stuck on the "Smelt ammy" step since that was always true