fix:Remove isVisible consideration in npcChanged for NpcStep
It's possible this will break a few highlight options in older quests (Mountain Daughter Kendal), but will need testing to see.

isVisible seems very hit-or-miss on actually stating if the NPC is visible or not. Even when checking post-spawn, it will still state invisible in the composition and transformed composition, despite being visible.