Sheep Shearer Rewrite (#1156)
* Sheep shearer re-write

* Fix wool check to work for start of Sheep Shearer

* Small improvements to Sheep Shearer

* Add in check for logged in checks in QuestHelperPlugin

Ensures it only runs on login/hop, not when the map is loading and other such actions.

* fix:Remove isVisible consideration in npcChanged for NpcStep

It's possible this will break a few highlight options in older quests (Mountain Daughter Kendal), but will need testing to see.

isVisible seems very hit-or-miss on actually stating if the NPC is visible or not. Even when checking post-spawn, it will still state invisible in the composition and transformed composition, despite being visible.

* fix:Make use of RunescapeProfileChanged

This should mean we only re-check requirements, bank items, and fake NPCs on initial login to a profile, and not in other scenarios such as hopping.