feat: Fire off a warning if a VarbitRequirement varbit read fails
This can happen if the varbit we requested no longer exists

If the varbit no longer exists, a warning will be posted to the users client.log file and the chat box. The step itself
will continue working but most likely not as intended, as the check function will return false.