Based of feedback from Pajlada
Still missing two pars below that I didn't want to do without being able to test but Zoinkwiz said he could look at it.

The same step should be split up into two steps
Varbit 22: "Head to Varrock Museum and speak to..."
Varbit 24: "Pickpocket Curator Haig..."

"Open the door to the room in the north east..." - since it's a "single use" door, you could highlight the door when the user is outside the orom, and highlight the crate when the user is inside the room. The text is small enough that it can stay the same for both steps.

Placement in optimal quest guide & optimal ironman quest guide. We do this one after all 4 quests are done or that wiki updates it.