Ethically Aquired Antiquities (#1767)
* Ethically Aquired Antiquities

This adds the new Novice quest of Ethically Aquired Antiquities, I need someone to test it and also some changes that I already sent to Zoinkwiz to fix.

* Ethically Aquired Antiquities extra fixes

Fixed some more issues to it

* Based of feedback from Pajlada

Still missing two pars below that I didn't want to do without being able to test but Zoinkwiz said he could look at it.

The same step should be split up into two steps
Varbit 22: "Head to Varrock Museum and speak to..."
Varbit 24: "Pickpocket Curator Haig..."

"Open the door to the room in the north east..." - since it's a "single use" door, you could highlight the door when the user is outside the orom, and highlight the crate when the user is inside the room. The text is small enough that it can stay the same for both steps.

Placement in optimal quest guide & optimal ironman quest guide. We do this one after all 4 quests are done or that wiki updates it.

* Wiki attempt at fixing (maybe?)

fixed links hopefully

* nevermind, one more

fix to the wiki link again

* fix: Formatting changes, and extra checks for sails

* fix: Hide sails from overlay req when given over

* fix: Add missing pickpocket step in EAA

* nit: Add differentiation between two same-phrased steps