While Guthix Sleeps feedback changes
Reworded the Item requirement of khazard teleport to show as
"2x Khazard teleports, Ardougne cloak or Spirit tree (3)"

Instead of
"2x Spirit tree (3), Ardougne cloak or Khazard teleport"

As this might look a bit confusing with the 2x spirit tree and the 3 digit after which is well the number, just looks more easily understandable this way.

Also changed the widget highlight to 44 instead of 47, as 47 is Smoggy. Issue was posted about it in
https://github.com/Zoinkwiz/quest-helper/issues/1821