fix: Avoid assist level panel being overriden by list on search change
When you had a search for a quest, you would have the assist panel shown it an assist level wasn't selected, but then the search text being set to empty would then cause the quest list panel to be made active, hiding it again.

This ensures that the active state of the assist panel is considered for trying to show the quest list.