fix: Allow for better handling for copied instances (#2007)
* fix: Allow for better handling for copied instances

This specifically came up in A Forgettable Tale... where the platforms with chests in the puzzle section exist a few times in the same space.

This meant sometimes the first LocalPoint found was actually the incorrect point. This change is intended to make it so that all LocalPoints can be considered by Steps before actually drawing arrows and highlights.

A note is I feel that the final section of DS2 might possibly be impacted by this, but I can't really predict in what way.

* fix: Ensure return from getInstanceLocalPointFromReal doesn't contain null values