feat: Added sidebar highlight depth toggle for ConditionalStep (#2078)
Sometimes we want to use a ConditionalStep as a sidebar step, but also sometimes in another spot just a normal step which is a step within a ConditionalStep.

This usually results in the sidebar highlighting both, as the step can be active in a depth search, as well as as the higher conditionalstep being active.

This allows for full depth searching to not be done for sidebar consideration for specific conditional steps.