Fix typo in Death on the Isle
The first step under the final "The show must go on" section had the word "theatre" an extra time. This change removes the redundant use of the word.

The text for this step could be more descriptive by mentioning the agility rocks to the south that have to be climbed first. It looks like there's code to do this but it wasn't functional when I went through the quest.