Reduce redundant helper text in Enakhra's Lament
Near the end of the "Craft a statue" section the text between two steps is redundant. The programmed logic works fine but the redundant text can cause confusion. This change removes the redundant text from the first step because the following step effectively guides the player on what to do next. This change also formats the weight of the granite in parentheses to match the name of the item.