unrelated: make header a JTextPane
this allows sections to span multiple lines which is cool for the Scrambled! quest imo but can be skipped if you want to
since it _does_ look a bit different. I don't know how to grapple Swing into doing proper vertical alignment

Slightly reduce bottom border of queststeppanel header

this makes it look more similar to when we used a JLabel