Fremennik Isles: Go to next step after getting Jester outfit (#2249)
Before this change, nothing happens after you get the Jester outfit from the chest. Only after equipping the Jester
outfit does it show the next step, but this is not explained to the player.

The `hasJesterOutfit` `ItemRequirement` now does not require the jester outfit to be equipped, and so now the next steps
will show correctly even when the Jester outfit is in your inventory.