fix: One Small Favour
One of the first steps of the quest is to talk to a Jungle Forester
south of Shilo to receive his blunt axe to be sharpened. However, the
talkToJungleForestStep incorrectly listed the blunt axe as a
requirement. This change removes that requirement (it's still correctly
a requirement for the talkToBrian step).