Feat multitier panels v2 (#2302)
* feat: Initial implementation of TopLevelPanelDetails

The intention is to allow for sub-sections of sections. The core goal is to allow for sub-ordering of parts of helpers, but also may be a nice QoL for allowing for larger more complex helpers which want to divide up into further chunks.

* Updated PanelDetails id to default to -1

* feat: Added TopLevelSectionGroup

This should group around a top level section.

* Added default constructor for an empty Top Level Panel Group

* feat: Make use of top level groups in side panel

This makes the side panel groupings visually distinct, whilst not disrupting the original look too much.

* Made top level sections more subtle in the sidebar

* fix: Hide wait for herbs step when not needed in Herb Run helper

* fix: Corrected allCollapsed check to still be against all sub panels

* feat: Make use of existing UI design for nesting sidebar panels

* feat: Update herb run and DT2 to use new nested sidebar

* feedback: multitier panels (#2319)

* nit: get rid of unneccessary getters

* nit: remove dupe defined members

* nit: unabstract getStepsLabel

* fix: PR feedback fixes

---------

Co-authored-by: pajlada <rasmus.karlsson@pajlada.com>