fix: add additional information to cabin fever plunder step (#2371)
When you have to loot the chest, crate and barrel during cabin fever, you only get 6 plunder, but you need 10 in total, which makes this step a bit confusing. According to the wiki, it can take up to 5 minutes for the plunder to respawn. I therefore think it would be helpful to tell the player to switch worlds for the loot to respawn instantly.

Source: https://oldschool.runescape.wiki/w/Cabin_Fever#Retaliation!